requirements.txt
langchain
weaviate-client
torch 
sentence-transformers 
langchain_cohere 
langchain-weaviate
pymupdf
langchain_openai
langchain-google-genai
streamlit
google-search-results